Added the main.c file
For this file, when you compile and run, the scanner will ask for a state abbreviation. Type in IA for Iowa. Potentially, whatever 2 characters the user types in, it could load all of the weather stations for them to pick from. In the HTTP get request, it says "state=ia". The IA part can be replaced with any state abbreviation and it can load all of the weather stations for the user to pick. It's not fully correct or even partial but it works. Maybe this can get us started with it.